+++ src/utils/Constants.js
@@ -9,6 +9,7 @@ define(function (require, exports, module) {
         INSERT_BTN: "button#zotero-insert-btn",
         CLEAR_BTN: "button#zotero-clear-btn",
         CLOSE_BTN: ".close",
+        SETTINGS_BTN: "i#zotero-settings-btn",
         INSERT_BIBLIO_BTN: "button#zotero-insert-biblio-btn",
         GENERATE_BIBLIO_BTN: "button#zotero-generate-biblio-file-btn",
 
@@ -20,6 +21,7 @@ define(function (require, exports, module) {
         CMD_ID_HIDE_PANEL: "baig.zoteroplugin.panel.hide",
         CMD_ID_TOGGLE_PANEL: "baig.zoteroplugin.panel.toggle",
         CMD_ID_CLEAR_ALL: "baig.zoteroplugin.clear.all",
+        CMD_ID_SHOW_SETTINGS: "baig.zoteroplugin.settings.show",
         CMD_ID_INSERT_CITE: "baig.zoteroplugin.citation.insert",
         CMD_ID_INSERT_BIBLIO: "baig.zoteroplugin.bibliography.insert",
         CMD_ID_GENERATE_BIBLIO: "baig.zoteroplugin.bibliography.generate",